requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dill==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.10.0
pycodestyle==2.11.0
PyJWT==1.7.1
pylint==2.17.5
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.1
tzdata==2023.3
urllib3==2.0.5
wrapt==1.15.0
